split out ecs back into object, component and system
