+++ include/framework/ecs/Component.h
@@ -1,5 +1,5 @@
-#ifndef HESTIA_FRAMEWORK_ECS_ECS_H
-#define HESTIA_FRAMEWORK_ECS_ECS_H
+#ifndef HESTIA_FRAMEWORK_ECS_COMPONENT_H
+#define HESTIA_FRAMEWORK_ECS_COMPONENT_H
 
 #include <vector>
 #include <string>
@@ -10,27 +10,6 @@
 
 namespace HGE {
     using UID = int;
-
-    /**
-     * Object Interface
-     */
-    class Object {
-
-        protected:
-        UID id = GenerateUniqueId();
-
-        public:
-        virtual ~Object() = default;
-
-        /* Setters and Getters */
-        UID getId() const {
-        return id;
-        }
-
-        /* Public Methods */
-        virtual void onCreate() = 0;
-        virtual void tick(double deltaTime) = 0;
-    };
 
     /**
      * Component Interface
@@ -70,27 +49,6 @@ namespace HGE {
         ComponentArray& operator= (const ComponentArray &other) = delete;
 
         std::vector<std::unique_ptr<C>>& getComponents() { return mComponents; }
-    };
-
-    /**
-     * System Interface
-     */
-    class ISystem {
-        public:
-        virtual void run() = 0;
-        virtual ~ISystem() = default;
-    };
-
-    /**
-     * System Template
-     */
-    template <class C>
-    class System : public ISystem {
-
-        System() = default;
-        public:
-        ~System() = default;
-        void run() = 0;
     };
 }
